Tambahkan konfigurasi VITE_API2_BASE_URL, sederhanakan service praikiraan-cuaca
