+++ frontend/.env.example
@@ -1,5 +1,6 @@
 #VITE_BASE_URL=https://bbwssumatera8.com
 VITE_BASE_URL=https://localhost/balai/bbwssumatera8/api
+VITE_API2_BASE_URL=http://localhost:3000/balai/bbwssumatera8/api2
 APP_ENV=development
 FRONTEND_PORT_MAPPING=127.0.0.1:5173:8008
 VITE_API_TARGET=https://localhost/balai/bbwssumatera8/api
